+++ Projekt/eeBF/src/Model/User.java
@@ -6,7 +6,7 @@ import java.util.Calendar;
 
 public abstract class User
 {
-	long id;
+	int id;
 	private String email;
 	private String passwort; //Encryption!
 	Calendar createdate;
@@ -19,12 +19,12 @@ public abstract class User
 		setPasswort(passwort);
 	}
 
-	public long getId() 
+	public int getId() 
 	{
 		return id;
 	}
 	
-	public void setId(long id)
+	public void setId(int id)
 	{
 		this.id = id;
 	}
